+++ extensions/arc/src/release-notes.tsx
@@ -1,7 +1,7 @@
 import { closeMainWindow, showHUD } from "@raycast/api";
 import { makeNewTab } from "./arc";
 
-const url = "https://resources.arc.net/en/articles/8233343-release-notes";
+const url = "https://resources.arc.net/hc/en-us/articles/20498285812375-Release-Notes";
 
 export default async function command() {
   try {