+++ extensions/tidyread---streamline-your-daily-reading/src/digest/index.ts
@@ -2,47 +2,24 @@ import Parser from "rss-parser";
 import pLimit from "p-limit";
 import { HttpsProxyAgent } from "https-proxy-agent";
 import { load } from "cheerio";
-import { isXML, normalizeUrlForMarkdown, retry, withTimeout } from "../utils/util";
-import { DigestItem, Provider, SummarizeStatus } from "../types";
+import { isXML, normalizeUrlForMarkdown, retry, sleep, withTimeout } from "../utils/util";
+import {
+  DigestItem,
+  DigestStage,
+  Provider,
+  PullItemsStage,
+  RSSFeed,
+  RSSItem,
+  SummarizeItemStage,
+  SummarizeStatus,
+} from "../types";
 import dayjs from "dayjs";
 import { uniqBy } from "lodash";
 import { addUtmSourceToUrl } from "../utils/biz";
 // import { fetchMetadata } from "../utils/request";
 
-export interface Preferences {
-  provider?: "openai";
-  apiKey?: string;
-  apiModel?: string;
-  apiHost?: string;
-  httpProxy?: string;
-  summarizePrompt?: string;
-  maxItemsPerFeed?: number;
-  maxApiConcurrency?: number;
-  notificationTime?: string;
-  autoGenDigest?: boolean;
-}
-
-export interface RSSFeed {
-  title: string;
-  url: string;
-  filter?: (item: RSSItem) => boolean;
-  maxItems?: number;
-}
-
-export interface RSSItem {
-  link?: string;
-  title?: string;
-  pubDate?: string | number;
-  creator?: string;
-  summary?: string;
-  content?: string;
-  coverImage?: string;
-  feed?: RSSFeed;
-  [k: string]: any;
-}
-
 export type RSSItemWithStatus = RSSItem & {
-  status: SummarizeStatus;
+  status?: SummarizeStatus;
 };
 
 const MIN_SUMMARIZE_CHARACTER_LIMIT = 100;
@@ -53,6 +30,7 @@ async function getAllFilteredItems(
   rssFeeds: RSSFeed[],
   httpProxy?: string,
   requestTimeout?: number,
+  onProgress?: (stage: PullItemsStage, err?: Error) => void,
 ): Promise<RSSItem[]> {
   const agent = httpProxy ? new HttpsProxyAgent(httpProxy) : undefined;
 
@@ -63,10 +41,16 @@ async function getAllFilteredItems(
   });
   let allItems: RSSItem[] = [];
 
+  onProgress?.({
+    stageName: "pull_items",
+    status: "start",
+    data: null,
+  });
+
   for (const feed of rssFeeds) {
     try {
       console.log("start to parse rss feed", feed.url);
-      const resp = await withTimeout(parser.parseURL(feed.url), requestTimeout ?? 20000);
+      const resp = await withTimeout(parser.parseURL(feed.url), requestTimeout ?? 30 * 1000);
 
       const filteredItems = await Promise.all(
         // 需要对相同title去重，有些rss质量不高，会出现重复的title不同link
@@ -90,9 +74,23 @@ async function getAllFilteredItems(
       allItems = allItems.concat(filteredItems);
     } catch (error: any) {
       console.error(`Failed to parse RSS feed ${feed.url}: ${error.message}`);
-      throw error;
+      onProgress?.(
+        {
+          stageName: "pull_items",
+          status: "failed",
+          data: null,
+        },
+        error,
+      );
+      throw new Error(`Failed to parse RSS feed ${feed.url}: ${error.message}`);
     }
   }
+
+  onProgress?.({
+    stageName: "pull_items",
+    status: "success",
+    data: allItems.length,
+  });
 
   return allItems;
 }
@@ -111,32 +109,63 @@ async function summarizeItem(
   requestTimeout?: number,
   retryCount?: number,
   retryDelay?: number,
+  onProgress?: (stage: SummarizeItemStage, err?: Error) => void,
 ): Promise<RSSItemWithStatus> {
-  console.time(`summarize call for ${item.title} @${item.feed?.title}`);
   // console.log(`retry count: ${retryCount}, retry delay: ${retryDelay}`);
+  const needSummarize = item.content && item.content.length > MIN_SUMMARIZE_CHARACTER_LIMIT && provider.available;
+
   try {
-    const needSummarize = item.content && item.content.length > MIN_SUMMARIZE_CHARACTER_LIMIT && provider.available;
+    onProgress?.({
+      stageName: "summarize_item",
+      status: "start",
+      data: item,
+      type: needSummarize ? "ai" : "raw",
+    });
+
     const summary = needSummarize
       ? await retry(
-          () => withTimeout(provider.summarize(item.content!), requestTimeout ?? 20000),
+          () => withTimeout(provider.summarize(item.content!), requestTimeout ?? 30 * 1000),
           retryCount ?? 5,
           retryDelay ?? 30 * 1000,
+          (err) => {
+            // moonshot会出现high risk的错误，这种错误不需要重试
+            if (err.message.includes("high risk")) {
+              return true;
+            }
+
+            return false;
+          },
         )
       : ellipsisContent(item.content || "", THRESHOLDS_FOR_TRUNCATION);
 
-    return { ...item, summary: summary, status: needSummarize ? "summraized" : "raw" };
+    onProgress?.({
+      stageName: "summarize_item",
+      status: "success",
+      data: item,
+      type: needSummarize ? "ai" : "raw",
+    });
+
+    // 如果开新对象，外部并发的titles赋值会无效
+    return Object.assign(item, {
+      summary: summary,
+      status: needSummarize ? ("summraized" as const) : ("raw" as const),
+    });
   } catch (error: any) {
     console.error(`Failed to summarize: ${error.message}`);
-    return {
-      ...item,
-      summary: `> **Fail to summarize**, error is: \`${error.message}\`. Raw content is below:\n\n${ellipsisContent(
-        item.content || "",
-        THRESHOLDS_FOR_TRUNCATION,
-      )}`,
-      status: "failedToSummarize",
-    };
-  } finally {
-    console.timeEnd(`summarize call for ${item.title} @${item?.feed?.title}`);
+    onProgress?.({
+      stageName: "summarize_item",
+      status: "failed",
+      data: item,
+      type: needSummarize ? "ai" : "raw",
+    });
+
+    // 如果开新对象，外部并发的titles赋值会无效
+    return Object.assign(item, {
+      summary: `> ❗ **Failed to summarize**, error is: \`${
+        error.message
+      }\`. Raw content is below:\n\n${ellipsisContent(item.content || "", THRESHOLDS_FOR_TRUNCATION)}`,
+      status: "failedToSummarize" as const,
+    });
   }
 }
 
@@ -145,13 +174,13 @@ function generateDigestTemplate(provider: Provider, items: RSSItemWithStatus[]):
   // const prefix = `# ${title}  \`at ${dayjs(time).format('HH:mm')}\`\n\n`;
   const prefix = provider.available
     ? ``
-    : `> **Your AI Provider has not been configured correctly**. When it is configured, each item will be summarized by AI, otherwise it will only get the original excerpts.\n\n`;
+    : `> 💡 **Your AI Provider has not been configured correctly**. When it is configured, each item will be summarized by AI, otherwise it will only get the raw content. Check [the doc](https://www.tidyread.info/docs/empowered-with-ai) to learn how to config.\n\n`;
   let digest = `${prefix}`;
 
-  digest += `## Introduction\n[Tidyread](https://tidyread.info) generated a flat summary of the content from all the sources today. **Only sources that have a valid RSS Link** can be summarized.\nThe content pulled by rss will be filtered according to the \`Time Span\` you provide, keeping only the content in the time span.\n\n## Summary\n`;
+  digest += `## Introduction\nTidyread generated a flat summary of the content from all the sources today. **Only sources that have a valid [RSS](https://meganesulli.com/blog/how-rss-works/) Link** can be summarized. Check [the doc](https://www.tidyread.info/docs/where-to-find-rss) to know where to find RSS.\n\n## Summary\n`;
 
   if (items.length === 0) {
-    return `${digest}No RSS items remain after filtering.`;
+    return `${digest}No [RSS](https://meganesulli.com/blog/how-rss-works/) items remain after filtering.`;
   }
 
   for (const [index, item] of items.entries()) {
@@ -159,21 +188,28 @@ function generateDigestTemplate(provider: Provider, items: RSSItemWithStatus[]):
   }
 
   if (items.some((item) => item.status === "failedToSummarize")) {
-    digest += `---\n\n## Why Some Articles Failed To Be Summarized By AI?\n\nYou can check out [this document](https://www.tidyread.info/docs/why-some-articles-fail-to-be-summarized) to understand why and how to fix it.\n\n`;
+    digest += `\n\n---\n\n### 🚧 Why Some Articles Failed To Be Summarized By AI?\n\nYou can check out [this document](https://www.tidyread.info/docs/why-some-articles-fail-to-be-summarized) to understand why and how to fix it.\n\n`;
   }
 
   return digest;
 }
 
+function addEllipsis(content: string, maxLen: number): string {
+  if (content.length > maxLen) {
+    return content.substring(0, maxLen) + "...";
+  }
+  return content;
+}
+
 // 格式化单个项目以用于摘要
-function formatItemForDigest(item: RSSItemWithStatus, prefixStr?: string): string {
-  return `### ${prefixStr ?? ""}${item.title}  @([${item?.feed?.title}](${addUtmSourceToUrl(
-    item?.feed?.url ?? "",
-  )}))\n${item.coverImage ? `![cover](${item.coverImage})\n\n` : ""}${
-    item.summary || item.content?.substring(0, THRESHOLDS_FOR_TRUNCATION) + "..."
+export function formatItemForDigest(item: RSSItemWithStatus, prefixStr?: string, ignoreFeed?: boolean): string {
+  return `### ${prefixStr ?? ""}${item.title}  ${
+    ignoreFeed ? "" : `@([${item?.feed?.title}](${addUtmSourceToUrl(item?.feed?.url ?? "")}))`
+  }\n${item.coverImage ? `![cover](${item.coverImage})\n\n` : ""}${
+    item.summary || addEllipsis(item.content || "", THRESHOLDS_FOR_TRUNCATION)
   }\n\n[Source Link](${normalizeUrlForMarkdown(item.link ?? "")})\n\n${
-    item.status === "summraized" ? `\`AI Summarized\`  ` : ""
-  }${["raw", "failedToSummarize"].includes(item.status) ? `\`Raw Content\`  ` : ""}\`Pub Time: ${dayjs(
+    item.status === "summraized" ? `\`✨AI Summarized\`  ` : ""
+  }${["raw", "failedToSummarize"].includes(item?.status ?? "") ? `\`Raw Content\`  ` : ""}\`Pub Date: ${dayjs(
     item.pubDate,
   ).format("YYYY-MM-DD HH:mm")}\`  \`Creator: ${item.creator ?? "none"}\`\n\n`;
 }
@@ -189,7 +225,7 @@ function extractTextFromXML(xml: string): string {
 }
 
 // 格式化RSS Items，对content进行判断，如果是xml，则提取其中text
-async function formatRSSItems(items: RSSItem[]): Promise<RSSItem[]> {
+export async function formatRSSItems(items: RSSItem[]): Promise<RSSItem[]> {
   return Promise.all(
     items.map(async (item) => {
       const content = (item.content || "").replace("&lt;", "<").replace("&gt;", ">");
@@ -219,6 +255,7 @@ export async function genDigest(options: {
   retryCount?: number;
   retryDelay?: number;
   itemLinkFormat?: (link: string, item: RSSItem) => string;
+  onProgress?: (stage: DigestStage, err?: Error) => void;
 }): Promise<{
   content: string;
   items: DigestItem[];
@@ -227,11 +264,12 @@ export async function genDigest(options: {
   const now = Date.now();
   console.time(`gen digest ${now}`);
   const limit = pLimit(options.maxApiConcurrency ?? 3);
+  const { onProgress } = options;
 
   const rssFeeds = options.rssFeeds;
 
   // 第一步：获取并过滤所有RSS items
-  const allFilteredItems = await getAllFilteredItems(rssFeeds, options.httpProxy, options.requestTimeout);
+  const allFilteredItems = await getAllFilteredItems(rssFeeds, options.httpProxy, options.requestTimeout, onProgress);
 
   // 第二步：格式化rss item
   let formatedItems = (await formatRSSItems(allFilteredItems)) as RSSItemWithStatus[];
@@ -239,26 +277,61 @@ export async function genDigest(options: {
   console.log(`${formatedItems.length} rss items found after filter`, formatedItems);
 
   // 第三步：翻译titles
-  if (options.translateTitles) {
-    try {
-      const translatedTitles = await options.translateTitles(formatedItems.map((item) => item.title || ""));
+  async function translateTitles() {
+    if (options.translateTitles) {
+      console.time("translate titles");
+      try {
+        onProgress?.({
+          stageName: "translate_titles",
+          status: "start",
+          data: null,
+        });
+        const translatedTitles = await options.translateTitles(formatedItems.map((item) => item.title || ""));
 
-      formatedItems.forEach((item, index) => {
-        item.title = translatedTitles[index] ?? item.title;
-      });
-    } catch (err: any) {
-      console.error("translate titles failed", err);
+        console.log("translated titles success:", translatedTitles);
+
+        formatedItems.forEach((item, index) => {
+          item.title = translatedTitles[index] ?? item.title;
+        });
+        onProgress?.({
+          stageName: "translate_titles",
+          status: "success",
+          data: null,
+        });
+      } catch (err: any) {
+        onProgress?.(
+          {
+            stageName: "translate_titles",
+            status: "failed",
+            data: null,
+          },
+          err,
+        );
+        console.error("translate titles failed", err);
+      } finally {
+        console.timeEnd("translate titles");
+      }
     }
   }
 
   // 第四步：并发地对items进行概述
-  const summarizedItems = await Promise.all(
-    formatedItems.map((item) =>
-      limit(() =>
-        summarizeItem(item, options.provider, options.requestTimeout, options.retryCount, options.retryDelay),
-      ),
+  const [, ...summarizedItems] = await Promise.all([
+    translateTitles(),
+    ...formatedItems.map((item) =>
+      limit(async () => {
+        // 避免translateTitles失败
+        await sleep(100);
+        return summarizeItem(
+          item,
+          options.provider,
+          options.requestTimeout,
+          options.retryCount,
+          options.retryDelay,
+          onProgress,
+        );
+      }),
     ),
-  );
+  ]);
 
   formatedItems = summarizedItems.map((item) => ({
     ...item,
@@ -271,7 +344,7 @@ export async function genDigest(options: {
   return {
     content: generateDigestTemplate(options.provider, formatedItems),
     items: formatedItems.map((item) => ({
-      status: item.status,
+      status: item.status!,
     })),
     createAt: Date.now(),
   };