+++ extensions/chatgpt/CHANGELOG.md
@@ -1,6 +1,17 @@
 # Changelog
 
-## [Fix] - 2024-1-19
+## [Feat] - 2024-01-31
+
+- Feature: Make responses display above the question
+
+## [Fix] - 2024-01-28
+
+- Fix: Proxy not working
+- Fix: Fetching model data lag entered the form view
+- Fix: Switching from "Ask Question" to "Full Text Input", the model change didn't work
+- Chore: Add react hook lint
+
+## [Fix] - 2024-01-19
 
 - Fix: Azure API 401 error