+++ extensions/tidyread---streamline-your-daily-reading/CHANGELOG.md
@@ -1,8 +1,28 @@
 # TidyRead - Streamline Your Daily Reading Changelog
 
+## [Add Search RSS Page] - 2024-01-30
+- add search rss page
+- add generate digest in background
+- add redirect route to improve onboarding experience
+- add some feedback iamges in degest
+- improve gen digest panel
+- change metadata images
+- fix default maxItemsPerFeed 10 => 5
+
+
+## [Improving Digest Speed] - 2024-01-26
+
+- add progress display for generating digest
+- openai api change to normal response instead of stream
+- improve translate prompt
+- change api config based on provider instead of same
+- translate and summarize in parallel
+- requestTimeout default to 30s
+
 ## [Onboarding Page] - 2024-01-24
+
 - add quickly generating sources by user interest
-- add retryCount and retryDelay setting to decrease article summarizing error 
+- add retryCount and retryDelay setting to decrease article summarizing error
 - improve speed of batch import sources
 - improve error handling and error message
 - add Delete All Sources
@@ -10,4 +30,5 @@
 - fix fetch large html content memory leak
 
 ## [Initial Version] - 2024-01-06
+
 init